[ C106 ] Apps - all handlers check app_secret

+++ enki/handlersrestapi.py
@@ -135,12 +135,16 @@ class HandlerAPIv1Logout( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
-			if user_id and auth_token:
-				if EnkiModelRestAPITokenVerify.delete_by_user_id_token( user_id, auth_token ):
-					success = True
-					error = ''
+			app_secret = jsonobject.get( 'app_secret', '')
+			if user_id and auth_token and app_secret:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					if EnkiModelRestAPITokenVerify.delete_by_user_id_token( user_id, auth_token ):
+						success = True
+						error = ''
+					else:
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -156,17 +160,21 @@ class HandlerAPIv1AuthValidate( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
-			if user_id and auth_token:
-				if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
-					user_displayname = enki.libdisplayname.get_display_name( user_id )
-					if user_displayname:
-						answer.update({ 'user_displayname' : user_displayname })
-						success = True
-						error = ''
+			app_secret = jsonobject.get( 'app_secret', '')
+			if user_id and auth_token and app_secret:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
+						user_displayname = enki.libdisplayname.get_display_name( user_id )
+						if user_displayname:
+							answer.update({ 'user_displayname' : user_displayname })
+							success = True
+							error = ''
+						else:
+							error = 'Not found'
 					else:
-						error = 'Not found'
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -182,24 +190,28 @@ class HandlerAPIv1OwnsProducts( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
+			app_secret = jsonobject.get( 'app_secret', '')
 			products = jsonobject.get( 'products', '')
-			if user_id and auth_token:
-				if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
-					if products:   # check which products in the list are activated by the user and return them
-						list_entities = enki.libstore.fetch_EnkiProductKey_by_activator_products_list( user_id, products )
-					else:    # no product specified, return all products activated by the user
-						list_entities = enki.libstore.fetch_EnkiProductKey_by_activator( user_id )
-					if list_entities:
-						list_products = []
-						for i, item in enumerate( list_entities ):
-							list_products.append( item.product_name )
-						answer.update({ 'products_owned' : list_products })
-						success = True
-						error = ''
+			if user_id and auth_token and app_secret:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
+						if products:   # check which products in the list are activated by the user and return them
+							list_entities = enki.libstore.fetch_EnkiProductKey_by_activator_products_list( user_id, products )
+						else:    # no product specified, return all products activated by the user
+							list_entities = enki.libstore.fetch_EnkiProductKey_by_activator( user_id )
+						if list_entities:
+							list_products = []
+							for i, item in enumerate( list_entities ):
+								list_products.append( item.product_name )
+							answer.update({ 'products_owned' : list_products })
+							success = True
+							error = ''
+						else:
+							error = 'Not found'
 					else:
-						error = 'Not found'
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -215,17 +227,21 @@ class HandlerAPIv1Friends( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
-			if user_id and auth_token:
-				if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
-					friends = enki.libfriends.get_friends_user_id_display_name( user_id )
-					if friends:
-						answer.update({ 'friends' : friends })
-						success = True
-						error = ''
+			app_secret = jsonobject.get( 'app_secret', '')
+			if user_id and auth_token and app_secret:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					if EnkiModelRestAPITokenVerify.exist_by_user_id_token( user_id, auth_token ):
+						friends = enki.libfriends.get_friends_user_id_display_name( user_id )
+						if friends:
+							answer.update({ 'friends' : friends })
+							success = True
+							error = ''
+						else:
+							error = 'Not found'
 					else:
-						error = 'Not found'
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -241,6 +257,7 @@ class HandlerAPIv1DataStoreSet( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
+			app_secret = jsonobject.get( 'app_secret', '')
 			data_type = jsonobject.get( 'data_type', '')
 			data_id = jsonobject.get( 'data_id', '')
 			data_payload = jsonobject.get( 'data_payload' )
@@ -250,29 +267,32 @@ class HandlerAPIv1DataStoreSet( webapp2.RequestHandler ):
 				time_to_expiry = enki.librestapi.DATASTORE_NON_EXPIRING   # 100 years
 			time_expires = datetime.datetime.now() + datetime.timedelta( seconds = time_to_expiry )
 			read_access = jsonobject.get( 'read_access', '' )
-			if user_id and auth_token and data_type and data_id and data_payload and time_expires:
-				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
-				if token_valid:   # user is valid
-					# add optional calculated properties to the data payload
-					if 'calc_ip_addr' in data_payload:    # IP address of the request
-						remote_address = self.request.remote_addr
-						data_payload.update({ 'calc_ip_addr' : remote_address })
-					data_store = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
-					if data_store:  # update
-						data_store.data_payload = data_payload
-						data_store.time_expires = time_expires  # update the expiry time
-						if read_access: # if read_access is not specified, don't update it
-							data_store.read_access = read_access
-					else:   # create new
-						if not read_access: # if read_access is not specified, use the default value defined in the model
-							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires )
-						else:
-							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires, read_access = read_access )
-					data_store.put()
-					success = True
-					error = ''
+			if user_id and auth_token and app_secret and data_type and data_id and data_payload and time_expires:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
+					if token_valid:   # user is valid
+						# add optional calculated properties to the data payload
+						if 'calc_ip_addr' in data_payload:    # IP address of the request
+							remote_address = self.request.remote_addr
+							data_payload.update({ 'calc_ip_addr' : remote_address })
+						data_store = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
+						if data_store:  # update
+							data_store.data_payload = data_payload
+							data_store.time_expires = time_expires  # update the expiry time
+							if read_access: # if read_access is not specified, don't update it
+								data_store.read_access = read_access
+						else:   # create new
+							if not read_access: # if read_access is not specified, use the default value defined in the model
+								data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires )
+							else:
+								data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires, read_access = read_access )
+						data_store.put()
+						success = True
+						error = ''
+					else:
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -288,20 +308,24 @@ class HandlerAPIv1DataStoreGet( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
+			app_secret = jsonobject.get( 'app_secret', '')
 			data_type = jsonobject.get( 'data_type', '')
 			data_id = jsonobject.get( 'data_id', '')
-			if user_id and auth_token and data_type and data_id:
-				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
-				if token_valid:   # user is valid
-					data_store_item = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id_not_expired( user_id, token_valid.app_id, data_type, data_id )
-					if data_store_item:
-						answer.update({ 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires ) , 'read_access' : data_store_item.read_access, 'server_time' : int( time.time())})
-						success = True
-						error = ''
+			if user_id and auth_token and app_secret and data_type and data_id:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
+					if token_valid:   # user is valid
+						data_store_item = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id_not_expired( user_id, token_valid.app_id, data_type, data_id )
+						if data_store_item:
+							answer.update({ 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires ) , 'read_access' : data_store_item.read_access, 'server_time' : int( time.time())})
+							success = True
+							error = ''
+						else:
+							error = 'Not found'
 					else:
-						error = 'Not found'
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -317,34 +341,38 @@ class HandlerAPIv1DataStoreGetList( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', '' ))
 			auth_token = jsonobject.get( 'auth_token', '' )
+			app_secret = jsonobject.get( 'app_secret', '')
 			data_type = jsonobject.get( 'data_type', '' )
 			read_access = jsonobject.get( 'read_access', '' )
-			if user_id and auth_token and data_type and ( read_access == 'public' or read_access == 'private' or read_access == 'friends' ):
-				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
-				if token_valid:   # user is valid
-					error = 'Not found'
-					data_store_list = []
-					if read_access == 'public':   # returns all data with read-access "public"
-						data_store_list = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_app_id_data_type_read_access_not_expired( token_valid.app_id, data_type, read_access )
+			if user_id and auth_token and app_secret and data_type and ( read_access == 'public' or read_access == 'private' or read_access == 'friends' ):
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
+					if token_valid:   # user is valid
+						error = 'Not found'
+						data_store_list = []
+						if read_access == 'public':   # returns all data with read-access "public"
+							data_store_list = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_app_id_data_type_read_access_not_expired( token_valid.app_id, data_type, read_access )
+						else:
+							people_list = []
+							if read_access == 'private':    # returns all user's data with read-access "private"
+								people_list = [ user_id ]
+							elif read_access == 'friends':    # returns list of user's friends' data with friends' read_access "friends"
+								people_list = enki.libfriends.get_friends_user_id( user_id )    # get the user's friends' ids
+							if people_list:
+								for person_id in people_list:   # get each persons' data
+									data_store_list = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_read_access_not_expired( person_id, token_valid.app_id, data_type, read_access )
+						if data_store_list:
+							data_payloads = []
+							for data_store_item in data_store_list:
+								data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_id' : data_store_item.data_id, 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
+							if data_payloads:
+								answer.update({ 'data_payloads' : data_payloads, 'server_time' : int( time.time())})
+								success = True
+								error = ''
 					else:
-						people_list = []
-						if read_access == 'private':    # returns all user's data with read-access "private"
-							people_list = [ user_id ]
-						elif read_access == 'friends':    # returns list of user's friends' data with friends' read_access "friends"
-							people_list = enki.libfriends.get_friends_user_id( user_id )    # get the user's friends' ids
-						if people_list:
-							for person_id in people_list:   # get each persons' data
-								data_store_list = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_read_access_not_expired( person_id, token_valid.app_id, data_type, read_access )
-					if data_store_list:
-						data_payloads = []
-						for data_store_item in data_store_list:
-							data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_id' : data_store_item.data_id, 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
-						if data_payloads:
-							answer.update({ 'data_payloads' : data_payloads, 'server_time' : int( time.time())})
-							success = True
-							error = ''
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
@@ -360,17 +388,21 @@ class HandlerAPIv1DataStoreDel( webapp2.RequestHandler ):
 		if jsonobject:
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
+			app_secret = jsonobject.get( 'app_secret', '')
 			data_type = jsonobject.get( 'data_type', '')
 			data_id = jsonobject.get( 'data_id', '')
-			if user_id and auth_token and data_type and data_id:
-				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
-				if token_valid:   # user is valid
-					data_stores = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
-					ndb.delete_multi( data_stores )
-					success = True
-					error = ''
+			if user_id and auth_token and app_secret and data_type and data_id:
+				if enki.librestapi.check_secret( user_id, auth_token, app_secret ):
+					token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
+					if token_valid:   # user is valid
+						data_stores = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
+						ndb.delete_multi( data_stores )
+						success = True
+						error = ''
+					else:
+						error = 'Unauthorised user'
 				else:
-					error = 'Unauthorised'
+					error = 'Unauthorised app'
 		answer.update({ 'success' : success, 'error' : error })
 		self.response.headers[ 'Content-Type' ] = 'application/json'
 		self.response.write( json.dumps( answer, separators=(',',':') ))
